Added bridge for Thread
f
