+++ shared/src/androidMain/kotlin/com/jizoquval/chanBrowser/shared/viewModel/BoardsListViewModel.kt
@@ -1,10 +1,10 @@
 package com.jizoquval.chanBrowser.shared.viewModel
 
-import com.jizoquval.chanBrowser.shared.boardsList.BoardsListStore
-import com.jizoquval.chanBrowser.shared.boardsList.model.BoardListEvent
-import com.jizoquval.chanBrowser.shared.boardsList.model.BoardListModel
-import com.jizoquval.chanBrowser.shared.boardsList.model.eventToIntent
-import com.jizoquval.chanBrowser.shared.boardsList.model.stateToModel
+import com.jizoquval.chanBrowser.shared.features.boardsList.BoardsListStore
+import com.jizoquval.chanBrowser.shared.features.boardsList.model.BoardListEvent
+import com.jizoquval.chanBrowser.shared.features.boardsList.model.BoardListModel
+import com.jizoquval.chanBrowser.shared.features.boardsList.model.eventToIntent
+import com.jizoquval.chanBrowser.shared.features.boardsList.model.stateToModel
 
 class BoardsListViewModel(
     store: BoardsListStore
