--- shared/src/commonMain/kotlin/com/jizoquval/chanBrowser/shared/board/uiModel/BoardEvent.kt
@@ -1,3 +1,0 @@
-package com.jizoquval.chanBrowser.shared.board.uiModel
-
-sealed class BoardEvent
