--- shared/src/commonMain/kotlin/com/jizoquval/chanBrowser/shared/board/uiModel/BoardModel.kt
@@ -1,15 +1,0 @@
-package com.jizoquval.chanBrowser.shared.board.uiModel
-
-data class BoardModel(
-    val boardName: String,
-    val isProgress: Boolean,
-    val threadList: List<Thread>
-) {
-    data class Thread(
-        val id: Long,
-        val title: String,
-        val message: String,
-        val postsCount: Int,
-        val filesCount: Int
-    )
-}
