--- shared/src/commonMain/kotlin/com/jizoquval/chanBrowser/shared/boardsList/model/BoardListEvent.kt
@@ -1,6 +1,0 @@
-package com.jizoquval.chanBrowser.shared.boardsList.model
-
-sealed class BoardListEvent {
-    class SelectFavoriteBoard(val boardId: Long) : BoardListEvent()
-    class SelectBoard(val boardId: Long) : BoardListEvent()
-}
