+++ shared/src/commonMain/kotlin/com/jizoquval/chanBrowser/shared/cache/models/BoardDto.kt
