+++ shared/src/commonMain/kotlin/com/jizoquval/chanBrowser/shared/features/boardsList/model/BoardListEvent.kt
@@ -1,0 +1,6 @@
+package com.jizoquval.chanBrowser.shared.features.boardsList.model
+
+sealed class BoardListEvent {
+    class SelectFavoriteBoard(val boardId: Long) : BoardListEvent()
+    class SelectBoard(val boardId: Long) : BoardListEvent()
+}
