+++ shared/src/commonMain/kotlin/com/jizoquval/chanBrowser/shared/features/threadsList/model/BoardEvent.kt
@@ -1,0 +1,3 @@
+package com.jizoquval.chanBrowser.shared.features.threadsList.model
+
+sealed class BoardEvent
