+++ shared/src/commonMain/kotlin/com/jizoquval/chanBrowser/shared/features/threadsList/model/Mapper.kt
@@ -1,0 +1,19 @@
+package com.jizoquval.chanBrowser.shared.features.threadsList.model
+
+import com.jizoquval.chanBrowser.shared.features.threadsList.BoardStore
+
+fun stateToModel(state: BoardStore.State): BoardModel {
+    return BoardModel(
+        boardName = state.boardName,
+        isProgress = state.isProgress,
+        threadList = state.threads.map { thread ->
+            BoardModel.Thread(
+                id = thread.idOnChan,
+                title = thread.title,
+                message = thread.message,
+                postsCount = thread.postsCount,
+                filesCount = thread.filesCount
+            )
+        }
+    )
+}
