+++ shared/src/commonMain/kotlin/com/jizoquval/chanBrowser/shared/network/dvach/IDvachApi.kt
@@ -1,7 +1,7 @@
 package com.jizoquval.chanBrowser.shared.network.dvach
 
-import com.jizoquval.chanBrowser.shared.network.json.BoardJson
-import com.jizoquval.chanBrowser.shared.network.json.ThreadsListJson
+import com.jizoquval.chanBrowser.shared.network.dvach.json.BoardJson
+import com.jizoquval.chanBrowser.shared.network.dvach.json.ThreadsListJson
 
 interface IDvachApi {
