+++ shared/src/commonMain/kotlin/com/jizoquval/chanBrowser/shared/network/dvach/json/Icon.kt
@@ -1,0 +1,10 @@
+package com.jizoquval.chanBrowser.shared.network.dvach.json
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Icon(
+    val name: String,
+    val num: Int,
+    val url: String
+)
