--- shared/src/commonMain/kotlin/com/jizoquval/chanBrowser/shared/network/json/Icon.kt
@@ -1,10 +1,0 @@
-package com.jizoquval.chanBrowser.shared.network.json
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class Icon(
-    val name: String,
-    val num: Int,
-    val url: String
-)
